# Conflicts:
#	project/OsEngine/Market/Servers/SmartCom/SmartComServer.cs
#	project/OsEngine/OsData/NewSecurityUi.xaml.cs
#	project/OsEngine/bin/Debug/OsEngine.exe

+++ project/OsEngine/Charts/CandleChart/ChartCandleMaster.cs
@@ -42,19 +42,32 @@ namespace OsEngine.Charts.CandleChart
             _name = nameBoss + "ChartMaster";
             _startProgram = startProgram;
 
-            ChartCandle = new WinFormsChartPainter(nameBoss, startProgram);
+            if(_startProgram != StartProgram.IsOsOptimizer)
+            {
+                UpDateChartPainter();
+                Load();
+            }
+           
+            _canSave = true;
+        }
+
+        private void UpDateChartPainter()
+        {
+            if(ChartCandle != null)
+            {
+                ChartCandle.Delete();
+                ChartCandle.ChartClickEvent -= ChartCandle_ChartClickEvent;
+                ChartCandle.LogMessageEvent -= NewLogMessage;
+                ChartCandle.ClickToIndexEvent -= _chartCandle_ClickToIndexEvent;
+                ChartCandle.SizeAxisXChangeEvent -= ChartCandle_SizeAxisXChangeEvent;
+            }
+
+            ChartCandle = new WinFormsChartPainter(_name, _startProgram);
 
             ChartCandle.ChartClickEvent += ChartCandle_ChartClickEvent;
             ChartCandle.LogMessageEvent += NewLogMessage;
             ChartCandle.ClickToIndexEvent += _chartCandle_ClickToIndexEvent;
             ChartCandle.SizeAxisXChangeEvent += ChartCandle_SizeAxisXChangeEvent;
-
-            if(startProgram != StartProgram.IsOsOptimizer)
-            {
-                Load();
-            }
-           
-            _canSave = true;
         }
 
         /// <summary>
@@ -63,6 +76,10 @@ namespace OsEngine.Charts.CandleChart
         /// </summary>
         private void Load()
         {
+            if(_startProgram == StartProgram.IsOsOptimizer)
+            {
+                return;
+            }
             if (!File.Exists(@"Engine\" + Name + @".txt"))
             {
                 return;
@@ -389,26 +406,67 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
-                for (int i = 0; _indicators != null && i < _indicators.Count; i++)
+                if(_indicators != null)
                 {
-                    _indicators[i].Delete();
+                    for (int i = 0; _indicators != null && i < _indicators.Count; i++)
+                    {
+                        _indicators[i].NeadToReloadEvent -= indicator_NeadToReloadEvent;
+                        _indicators[i].Clear();
+
+                        if (_startProgram != StartProgram.IsOsOptimizer)
+                        {
+                            _indicators[i].Delete();
+                        }
+                    }
+
+                    _indicators.Clear();
+                    _indicators = null;
                 }
 
-                if (File.Exists(@"Engine\" + Name + @".txt"))
+                if(_startProgram != StartProgram.IsOsOptimizer)
                 {
-                    File.Delete(@"Engine\" + Name + @".txt");
+                    if (File.Exists(@"Engine\" + Name + @".txt"))
+                    {
+                        File.Delete(@"Engine\" + Name + @".txt");
+                    }
                 }
+
                 if (ChartCandle != null)
                 {
+                    ChartCandle.ChartClickEvent -= ChartCandle_ChartClickEvent;
+                    ChartCandle.ClickToIndexEvent -= _chartCandle_ClickToIndexEvent;
+                    ChartCandle.SizeAxisXChangeEvent -= ChartCandle_SizeAxisXChangeEvent;
+                    ChartCandle.ClearDataPointsAndSizeValue();
                     ChartCandle.Delete();
+                    ChartCandle.LogMessageEvent -= NewLogMessage;
+                    ChartCandle = null;
                 }
 
                 _myCandles = null;
-                _chartElements = null;
-                _alertArray = null;
-                _indicators = null;
-                _myPosition = null;
+
+                if (_chartElements != null)
+                {
+                    for(int i = 0;i < _chartElements.Count;i++)
+                    {
+                        _chartElements[i].UpdeteEvent -= myElement_UpdeteEvent;
+                        _chartElements[i].DeleteEvent -= myElement_DeleteEvent;
+                        _chartElements[i].Delete();
+                    }
+
+                    _chartElements.Clear();
+                    _chartElements = null;
+                }
+                if(_alertArray != null)
+                {
+                    _alertArray.Clear();
+                    _alertArray = null;
+                }
                 
+                if(_myPosition != null)
+                {
+                    _myPosition.Clear();
+                    _myPosition = null;
+                }
             }
             catch (Exception error)
             {
@@ -669,12 +727,6 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
-                if (((MenuItem)sender).Text == @"Trades")
-                {
-                    ChartCandle.DeleteTickArea();
-                    Save();
-                    return;
-                }
                 int number = ((MenuItem)sender).Index;
 
                 if ((_indicators == null || _indicators.Count <= number))
@@ -709,7 +761,7 @@ namespace OsEngine.Charts.CandleChart
             try
             {
                 IndicarotCreateUi ui = new IndicarotCreateUi(this);
-                ui.Show();
+                ui.ShowDialog();
 
                 if (IndicatorUpdateEvent != null)
                 {
@@ -798,6 +850,11 @@ namespace OsEngine.Charts.CandleChart
             {
                 indicator.NameArea = nameArea;
 
+                if (indicator.GetType().BaseType.Name == "Aindicator")
+                {
+                    ((Aindicator)indicator).StartProgram = _startProgram;
+                }
+
                 if (_indicators != null)
                 {
                     // check if there is such indicator in the collection
@@ -811,7 +868,20 @@ namespace OsEngine.Charts.CandleChart
                     }
                 }
 
+                if (_indicators == null)
+                {
+                    _indicators = new List<IIndicator>();
+                    _indicators.Add(indicator);
+                }
+                else
+                {
+                    _indicators.Add(indicator);
+                }
+
                 LoadIndicatorOnChart(indicator);
+
+                indicator.NeadToReloadEvent += indicator_NeadToReloadEvent;
+                indicator_NeadToReloadEvent(indicator);
 
                 return indicator;
             }
@@ -824,6 +894,10 @@ namespace OsEngine.Charts.CandleChart
 
         private void LoadIndicatorOnChart(IIndicator indicator)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             bool inNewArea = true;
             string nameArea = indicator.NameArea;
 
@@ -882,22 +956,8 @@ namespace OsEngine.Charts.CandleChart
                 }
             }
 
-
-
-            if (_indicators == null)
-            {
-                _indicators = new List<IIndicator>();
-                _indicators.Add(indicator);
-            }
-            else
-            {
-                _indicators.Add(indicator);
-            }
-
             Save();
             ReloadContext();
-            indicator.NeadToReloadEvent += indicator_NeadToReloadEvent;
-            indicator_NeadToReloadEvent(indicator);
         }
 
         /// <summary>
@@ -909,14 +969,16 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
-
                 List<Candle> candles = _myCandles;
                 if (candles != null)
                 {
                     indicator.Process(candles);
                 }
 
-                ChartCandle.RePaintIndicator(indicator);
+                if (ChartCandle != null)
+                {
+                    ChartCandle.RePaintIndicator(indicator);
+                }
             }
             catch (Exception error)
             {
@@ -931,9 +993,13 @@ namespace OsEngine.Charts.CandleChart
         /// <param name="indicator">indicator/индикатор</param>
         public void DeleteIndicator(IIndicator indicator)
         {
+
             try
             {
-                ChartCandle.DeleteIndicator(indicator);
+                if(ChartCandle != null)
+                {
+                    ChartCandle.DeleteIndicator(indicator);
+                }
 
                 indicator.Delete();
 
@@ -962,6 +1028,11 @@ namespace OsEngine.Charts.CandleChart
         /// <returns>true-created//true - создан//false-no//false - нет</returns>
         public bool IndicatorIsCreate(string name)
         {
+            if(ChartCandle == null)
+            {
+                return false;
+            }
+
             return ChartCandle.IndicatorIsCreate(name);
         }
 
@@ -973,7 +1044,14 @@ namespace OsEngine.Charts.CandleChart
         /// <returns></returns>
         public bool AreaIsCreate(string name)
         {
-            return ChartCandle.AreaIsCreate(name);
+            if(ChartCandle != null)
+            {
+                return ChartCandle.AreaIsCreate(name);
+            }
+            else
+            {
+                return true;
+            }
         }
         // chart element management
         // управление элементами чарта
@@ -993,10 +1071,17 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
+
                 if (_chartElements == null)
                 {
                     _chartElements = new List<IChartElement>();
                 }
+
+                if(_startProgram == StartProgram.IsOsOptimizer)
+                {
+                    return;
+                }
+
                 // 1  If such an element has already been added, replace it
                 // 1 если такой элемент уже добавлен, заменяем его
                 IChartElement myElement = _chartElements.Find(chartElement => chartElement.UniqName == element.UniqName);
@@ -1015,8 +1100,12 @@ namespace OsEngine.Charts.CandleChart
                 myElement.DeleteEvent += myElement_DeleteEvent;
                 // 2 sending it over for a drawing.
                 // 2 отправляем на прорисовку
-
-                ChartCandle.ProcessElem(myElement);
+                
+                if(ChartCandle != null && _startProgram 
+                    != StartProgram.IsOsOptimizer)
+                {
+                    ChartCandle.ProcessElem(myElement);
+                }
             }
             catch (Exception error)
             {
@@ -1031,8 +1120,11 @@ namespace OsEngine.Charts.CandleChart
         /// <param name="element">element/элемент</param>
         public void DeleteChartElement(IChartElement element)
         {
-            ChartCandle.ProcessClearElem(element);
-
+            if(ChartCandle != null)
+            {
+                ChartCandle.ProcessClearElem(element);
+            }
+            
             try
             {
                 // if there is such an element in the collection of elements - delete
@@ -1095,7 +1187,10 @@ namespace OsEngine.Charts.CandleChart
             {
                 _chartElements.Add(element);
             }
-            ChartCandle.ProcessElem(element);
+            if(ChartCandle != null)
+            {
+                ChartCandle.ProcessElem(element);
+            }
         }
       
         // Alert management управление Алертов
@@ -1127,7 +1222,10 @@ namespace OsEngine.Charts.CandleChart
                 {
                     if (alertArray[i].TypeAlert == AlertType.ChartAlert)
                     {
-                        ChartCandle.PaintAlert((AlertToChart)alertArray[i]);
+                        if(ChartCandle != null)
+                        {
+                            ChartCandle.PaintAlert((AlertToChart)alertArray[i]);
+                        }
                     }
                 }
             }
@@ -1146,7 +1244,10 @@ namespace OsEngine.Charts.CandleChart
         /// </summary>
         private void EraseAlertFromChart(List<IIAlert> alertArray)
         {
-            ChartCandle.ClearAlerts(alertArray);
+            if(ChartCandle != null)
+            {
+                ChartCandle.ClearAlerts(alertArray);
+            }
         }
 
         // candle drawing прорисовка свечей
@@ -1185,7 +1286,8 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
-                if (candles == null)
+                if (candles == null 
+                    || candles.Count == 0)
                 {
                     return;
                 }
@@ -1240,31 +1342,16 @@ namespace OsEngine.Charts.CandleChart
                         }
                     }
                 }
-
-            }
-            catch (Exception error)
-            {
-                SendErrorMessage(error);
-            }
-        }
-
-        // drawing ticks прорисовка тиков
-
-        /// <summary>
-        /// ticks in connector have been updated
-        /// в коннекторе обновились тики
-        /// </summary>
-        /// <param name="trades">ticks/тики</param>
-        public void SetTick(List<Trade> trades)
-        {
-            if(_startProgram == StartProgram.IsOsOptimizer)
-            {
-                return;
-            }
-
-            try
-            {
-                ChartCandle.ProcessTrades(trades);
+                else
+                {
+                    if (_indicators != null)
+                    {
+                        for (int i = 0; i < _indicators.Count; i++)
+                        {
+                            _indicators[i].Process(candles);
+                        }
+                    }
+                }
             }
             catch (Exception error)
             {
@@ -1305,6 +1392,10 @@ namespace OsEngine.Charts.CandleChart
         {
             try
             {
+                if(ChartCandle == null)
+                {
+                    return;
+                }
                 ChartCandle.StartPaintPrimeChart(gridChart, host, rectangle);
                 ChartCandle.ProcessCandles(_myCandles);
                 ChartCandle.ProcessPositions(_myPosition);
@@ -1341,13 +1432,18 @@ namespace OsEngine.Charts.CandleChart
         /// </summary>
         public void StopPaint()
         {
-            ChartCandle.StopPaint();
-
+            if (ChartCandle != null)
+            {
+                ChartCandle.StopPaint();
+            }
+           
             if (_grid != null)
             {
                 _grid.Children.Clear();
                 _grid = null;
             }
+
+            //UpDateChartPainter();
         }
 
         /// <summary>
@@ -1357,7 +1453,12 @@ namespace OsEngine.Charts.CandleChart
         public void Clear()
         {
             _myCandles = null;
-            ChartCandle.ClearDataPointsAndSizeValue();
+
+            if(ChartCandle != null)
+            {
+                ChartCandle.ClearDataPointsAndSizeValue();
+            }
+           
             _myPosition = null;
 
             for (int i = 0; _indicators != null && i < _indicators.Count; i++)
@@ -1368,7 +1469,10 @@ namespace OsEngine.Charts.CandleChart
 
         public void ClearTimePoints()
         {
-            ChartCandle.ClearDataPointsAndSizeValue();
+            if (ChartCandle != null)
+            {
+                ChartCandle.ClearDataPointsAndSizeValue();
+            }
         }
 
         /// <summary>
@@ -1376,21 +1480,37 @@ namespace OsEngine.Charts.CandleChart
         /// </summary>
         public void MoveChartToTheRight()
         {
-            ChartCandle.MoveChartToTheRight();
+            if (ChartCandle != null)
+            {
+                ChartCandle.MoveChartToTheRight();
+            }
         }
 
         public int GetSelectCandleNumber()
         {
+            if (ChartCandle == null)
+            {
+                return 0;
+            }
+
             return ChartCandle.GetCursorSelectCandleNumber();
         }
 
         public decimal GetCursorSelectPrice()
         {
+            if (ChartCandle == null)
+            {
+                return 0;
+            }
             return ChartCandle.GetCursorSelectPrice();
         }
 
         public void RemoveCursor()
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             ChartCandle.RemoveCursor();
         }
 
@@ -1401,11 +1521,19 @@ namespace OsEngine.Charts.CandleChart
         /// <param name="time">time/время</param>
         public void GoChartToTime(DateTime time)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             ChartCandle.GoChartToTime(time);
         }
 
         public void GoChartToIndex(int index)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             if (index < 0)
             {
                 return;
@@ -1420,6 +1548,10 @@ namespace OsEngine.Charts.CandleChart
         /// <returns>an array of areas. If not - null/массив областей. Если нет - null</returns>
         public List<string> GetChartAreas()
         {
+            if (ChartCandle == null)
+            {
+                return null;
+            }
             return ChartCandle.GetAreasNames();
         }
         // tool display window
@@ -1427,6 +1559,10 @@ namespace OsEngine.Charts.CandleChart
 
         public void StartPaintChartControlPanel(System.Windows.Controls.Grid grid)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             if (_label == null)
             {
                 _label = new System.Windows.Controls.Label();
@@ -1450,6 +1586,10 @@ namespace OsEngine.Charts.CandleChart
         /// <param name="serverType">server type/тип сервера</param>
         public void SetNewSecurity(string security, TimeFrameBuilder timeFrameBuilder, string portfolioName, ServerType serverType)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             if (_securityOnThisChart == security &&
                 _timeFrameSecurity == timeFrameBuilder.TimeFrame &&
                 serverType == _serverType &&
@@ -1520,6 +1660,10 @@ namespace OsEngine.Charts.CandleChart
         /// </summary>
         public string GetChartLabel()
         {
+            if (ChartCandle == null)
+            {
+                return "";
+            }
             string security = _securityOnThisChart;
 
             if (string.IsNullOrEmpty(security))
@@ -1544,6 +1688,11 @@ namespace OsEngine.Charts.CandleChart
         private void PaintLabelOnSlavePanel()
         {
             if (_grid == null)
+            {
+                return;
+            }
+
+            if (ChartCandle == null)
             {
                 return;
             }
@@ -1590,12 +1739,19 @@ namespace OsEngine.Charts.CandleChart
 
         public void PaintInDifColor(int indexStart, int indexEnd, string seriesName)
         {
+            if (ChartCandle == null)
+            {
+                return;
+            }
             ChartCandle.PaintInDifColor(indexStart, indexEnd, seriesName);
         }
 
         public void RefreshChartColor()
         {
-
+            if (ChartCandle == null)
+            {
+                return;
+            }
             ChartCandle.RefreshChartColor();
         }
